server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.8.2
requests==2.32.3
openai==1.40.0
pytesseract==0.3.10
Pillow==10.4.0
